feat(engine): consolidate editor UI, scene, and export infrastructure
Refactor the editor retained-host UI into focused app, callback, paint,
pointer, window, profiling, chrome command stream, and template-node modules.

Expand workbench behavior for asset panes, asset drag payloads, build/export
actions, module plugin actions, native window prompts, invalidation, viewport
toolbar routing, welcome sessions, workspace docking, and profiling output.

Restructure runtime dynamic scene support into folder-backed session, slot,
capture, restore, import/export, mutation, retention, validation, asset reload,
spawn task, and value modules.

Modularize ECS internals for archetypes, components, entities, commands,
resources, events, messages, observers, resource stores, and component storage.

Advance rendering infrastructure with camera history keys, render queue sort
keys, frame submission reports, scene clear and transparent passes, final
target output, light-grid stats, UI stats, product reports, and render region
coverage.

Improve export tooling with compile-host reporting, native dynamic build and
payload schemas, platform bundle template resolution, source template checks,
stage handoff metadata, operation audit reports, and expanded schema tests.

Update runtime and app integration for host request routing, profile/bootstrap
guards, dynamic API diagnostics, input cursor state, asset resource resolution,
profiling contracts, and plugin export materialization.

Refresh runtime, render, editor UI, export, scene, ECS, and architecture docs
to match the new module ownership, hard-cutover structure, diagnostics, and
plan status tables.

+++ zircon_runtime/src/core/framework/render/core_pipeline/phase_queue_ordering_key.rs
@@ -20,7 +20,7 @@ impl RenderPhaseQueueOrderingKey {
         }
     }
 
-    pub const fn raw_sort_key(self) -> i128 {
+    pub const fn raw_sort_key(self) -> u64 {
         self.sort_key.raw()
     }
 }
